List thumbnails horizontally
We want a list of thumbnails which are located next to each other. In order to get it, we set list items float.  Notice we changed html strocture in index.html. We put anchor tag in a div to get better contorl over spcae such as padding and margin around anchor tag. The main reason to this tweak is that <a> is clickable but <div> is not.  Thumbnails are 80px wide and 40px height and also every thumbnail get 5px border on its each side. Finally a thumbnail's total width would be 90px and its total height would be 50px.  The anchor contains an image and a span. We want to give a value to height and width property of anchors, so the inline-block property has been set for anchors. By considering height and width of thumbnails we set these values for <a> dimensions. By setting these dimension the overflow occures and there is no space in anchor for <span> to be placed there. We handle spans later. Notice we put widht of <li> equal to 14.285714286%. It helps to get responsive web page. Every li gets about 14.28 percent width.