Put triangle top of tooltip
Putin a triangle top of tooltip makes it perfect. We use a special
tecnique to create tirangle. We use ::before selector and create and
empty content element before <span>. Then we set border on created
element. But we set 3 side of borders transparent and another one equal
to our favorite color.